# back/swagger.yaml
openapi: 3.0.0
info:
  title: Products API
  description: API for managing products
  version: 1.0.0
components:
  schemas:
    Product:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: number
        inventoryStatus:
          type: string
        category:
          type: string
        image:
          type: string
        rating:
          type: number
paths:
  /api/products:
    get:
      description: Get all products
      responses:
        "200":
          description: Successful operation
        "500":
          description: Internal Server Error
    post:
      summary: Add a new product
      requestBody:
        description: Product details
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: Product added successfully.
        "500":
          description: Internal Server Error.
  /api/products/{id}:
    patch:
      summary: Update an existing product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the product to be updated.
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: Updated Product
              price: 29.99
      responses:
        "200":
          description: Successful operation. Returns the updated product.
        "500":
          description: Internal Server Error
    delete:
      summary: Delete a product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the product to be deleted.
      responses:
        "200":
          description: Product deleted successfully.
        "500":
          description: Internal Server Error.